@@ -37,19 +37,19 @@
     "del": "3.0.0",
     "elem-dataset": "1.1.1",
     "glob-all": "3.1.0",
-    "gulp": "3.9.1",
+    "gulp": "4.0.0",
     "gulp-bless": "4.0.0",
     "gulp-changed": "3.2.0",
     "gulp-clean-css": "3.9.3",
     "gulp-concat": "2.6.1",
     "gulp-eslint": "4.0.1",
-    "gulp-header": "1.8.9",
+    "gulp-header": "2.0.5",
     "gulp-imagemin": "4.1.0",
-    "gulp-less": "3.4.0",
+    "gulp-less": "4.0.0",
     "gulp-modernizr-build": "0.0.5",
     "gulp-mq-remove": "0.0.2",
     "gulp-newer": "1.4.0",
-    "gulp-notify": "3.1.0",
+    "gulp-notify": "3.2.0",
     "gulp-postcss": "7.0.1",
     "gulp-rename": "1.2.2",
     "gulp-replace": "0.6.1",
@@ -72,7 +72,7 @@
     "uglifyjs-webpack-plugin": "1.2.4",
     "ustream-embedapi": "1.0.0",
     "vinyl-named": "1.1.0",
-    "webpack": "4.3.0",
+    "webpack": "4.5.0",
     "webpack-stream": "4.0.3",
     "xdr": "0.5.3"
   },
@@ -95,6 +95,7 @@
     "protractor": "5.3.0",
     "protractor-accessibility-plugin": "github:cfpb/protractor-accessibility-plugin#on-page-load",
     "protractor-cucumber-framework": "4.2.0",
+    "protractor-retry": "1.1.9",
     "selenium-webdriver": "3.6.0",
     "snyk": "1.70.2",
     "wcag": "0.3.0"
