@@ -121,18 +121,33 @@ urlpatterns = [
         ),
         name='mortgage-estimate'
     ),
-    url(r'^owning-a-home/process/',
-        include(oah.urls_for_prefix('process/prepare/'))),
-    url(r'^owning-a-home/process/prepare/',
-        include(oah.urls_for_prefix('process/prepare/'))),
-    url(r'^owning-a-home/process/explore/',
-        include(oah.urls_for_prefix('process/explore/'))),
-    url(r'^owning-a-home/process/compare/',
-        include(oah.urls_for_prefix('process/compare/'))),
-    url(r'^owning-a-home/process/close/',
-        include(oah.urls_for_prefix('process/close/'))),
-    url(r'^owning-a-home/process/sources/',
-        include(oah.urls_for_prefix('process/sources/'))),
+
+    # Temporarily serve sheer version of OAH Journey sources page.
+    # TODO: remove once page is migrated into Wagtail.
+    flagged_url('OAH_JOURNEY_SHEER_SOURCE_PAGE',
+        r'^owning-a-home/process/sources(?P<path>.*)$',
+        lambda req, path: SheerTemplateView.as_view(
+            template_engine='owning-a-home',
+            template_name='process/sources{}index.html'.format(path or '/')
+        )(req),
+        fallback=lambda req, path: ServeView.as_view()(
+            req, 'owning-a-home/sources{}'.format(path or '/')
+        ),
+    ),
+    # END TODO
+
+    # Temporarily serve Wagtail OAH journey pages at `/process/` urls.
+    # TODO: change to redirects after 2018 homebuying campaign.
+    flagged_url('OAH_JOURNEY', r'^owning-a-home/process/(?P<path>.*)$',
+        lambda req, path: ServeView.as_view()(
+            req, 'owning-a-home/{}'.format(path or 'prepare/')
+        ),
+        fallback=lambda req, path: SheerTemplateView.as_view(
+            template_engine='owning-a-home',
+            template_name='process/{}/index.html'.format(path or 'prepare')
+        )(req)
+    ),
+    # END TODO
 
     url(r'^know-before-you-owe/$',
         TemplateView.as_view(
