@@ -1,16 +1,15 @@
+// TODO: Remove jquery.
 import $ from 'jquery';
+
+import {
+  analyticsLog,
+  delay,
+  track
+} from './util/analytics-util';
 
 // Paying for College custom analytics file
 
 const PFCAnalytics = ( function() {
-  // -- Delay calculations after keyup --//
-  const delay = ( function() {
-    let t = 0;
-    return function( callback, delay ) {
-      clearTimeout( t );
-      t = setTimeout( callback, delay );
-    };
-  } )(); // end delay()
 
   // -- findEmptyColumn() - finds the first empty column, returns column number [1-3] --//
   function findEmptyColumn() {
@@ -35,11 +34,7 @@ const PFCAnalytics = ( function() {
   $( '.remove-confirm .remove-yes' ).click( function() {
     const columnNumber = $( this ).parents( '[data-column]' ).attr( 'data-column' );
     const schoolID = $( '#institution-row [data-column="' + columnNumber + '"]' ).attr( 'data-schoolid' );
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'School Cost Comparison',
-      label: 'School Removed'
-    } );
+    track( 'School Interactions', 'School Cost Comparison', 'School Removed' );
     // Important to add a School tracking - reset the global.emptyColumn var
     global.emptyColumn = findEmptyColumn();
   } );
@@ -52,11 +47,7 @@ const PFCAnalytics = ( function() {
       const leftToPay = $( '.breakdown [data-column="' + columnNumber + '"] [data-nickname="gap"]' ).html();
       const schoolID = $( '#institution-row [data-column="' + columnNumber + '"]' ).attr( 'data-schoolid' );
       if ( leftToPay === '$0' && totalCosts !== '$0' ) {
-        window.dataLayer.push( {
-          event: 'School Interactions',
-          action: 'Reached Zero Left to Pay',
-          label: schoolID
-        } );
+        track( 'School Interactions', 'Reached Zero Left to Pay', schoolID );
       }
     }, 1000 );
   } );
@@ -64,11 +55,7 @@ const PFCAnalytics = ( function() {
   // Fire an event when a tooltip is clicked
   $( '.tooltip-info' ).click( function( event ) {
     const tooltip = $( this ).attr( 'data-tipname' );
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'Tooltip Clicked',
-      label: tooltip
-    } );
+    track( 'School Interactions', 'Tooltip Clicked', tooltip );
   } );
 
   // Fire an event when GI Bill panel opens
@@ -78,11 +65,7 @@ const PFCAnalytics = ( function() {
     delay( function() {
       const GIPanel = $( '[data-column="' + columnNumber + '"] .gibill-panel' );
       if ( GIPanel.is( ':visible' ) ) {
-        window.dataLayer.push( {
-          event: 'School Interactions',
-          action: 'GI Bill Calculator Opened',
-          label: schoolID
-        } );
+        track( 'School Interactions', 'GI Bill Calculator Opened', schoolID );
       }
     }, 500 );
   } );
@@ -92,11 +75,7 @@ const PFCAnalytics = ( function() {
     const buttonID = $( this ).attr( 'data-buttonid' );
     if ( $.inArray( buttonID, rateChangeClicks ) === -1 ) {
       rateChangeClicks.push( buttonID );
-      window.dataLayer.push( {
-        event: 'School Interactions',
-        action: 'Percent Arrow Clicked',
-        label: buttonID
-      } );
+      track( 'School Interactions', 'Percent Arrow Clicked', buttonID );
     }
 
   } );
@@ -110,119 +89,69 @@ const PFCAnalytics = ( function() {
     const residency = $( "[data-column='1'] .military-residency-panel :radio:checked" ).val();
     const control = $( '.header-cell[data-column="' + columnNumber + '"]' ).attr( 'data-control' );
 
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'GI Bill Calculator Submit',
-      label: schoolID
-    } );
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'Military Status',
-      label: serving
-    } );
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'Cumulative service',
-      label: tier
-    } );
+    track( 'School Interactions', 'GI Bill Calculator Submit', schoolID );
+    track( 'School Interactions', 'Military Status', serving );
+    track( 'School Interactions', 'Cumulative service', tier );
     if ( control === 'Public' ) {
-      window.dataLayer.push( {
-        event: 'School Interactions',
-        action: 'GI Residency',
-        label: residency
-      } );
+      track( 'School Interactions', 'GI Residency', residency );
     }
   } );
 
   // Fire an event when Send Email is clicked
   $( '#send-email' ).click( function() {
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'Save and Share',
-      label: 'Send email'
-    } );
+    track( 'School Interactions', 'Save and Share', 'Send email' );
   } );
 
   // Fire an event when save draw is opened
   $( '#save-and-share' ).click( function( event, nateeve ) {
     let UNDEFINED;
     if ( nateeve === UNDEFINED ) {
-      window.dataLayer.push( {
-        event: 'School Interactions',
-        action: 'Save and Share',
-        label: 'toggle button'
-      } );
+      track( 'School Interactions', 'Save and Share', 'toggle button' );
     }
   } );
 
   // Fire an event when save current is clicked
   $( '#save-current' ).click( function() {
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'Save and Share',
-      label: 'Save current worksheet'
-    } );
+    track( 'School Interactions', 'Save and Share', 'Save current worksheet' );
   } );
 
   $( '#unique' ).click( function() {
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'Save and Share',
-      label: 'Copy URL'
-    } );
+    track( 'School Interactions', 'Save and Share', 'Copy URL' );
   } );
 
   $( '#save-drawer .save-share-facebook' ).click( function() {
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'Save and Share',
-      label: 'Facebook_saveshare'
-    } );
+    track( 'School Interactions', 'Save and Share', 'Facebook_saveshare' );
   } );
 
   $( '#save-drawer .save-share-twitter' ).click( function() {
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'Save and Share',
-      label: 'Twitter_saveshare'
-    } );
+    track( 'School Interactions', 'Save and Share', 'Twitter_saveshare' );
   } );
 
   // Fire an event when Get Started is clicked
   $( '#get-started-button' ).click( function() {
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'School Cost Comparison',
-      label: 'Get Started Button'
-    } );
+    track(
+      'School Interactions', 'School Cost Comparison', 'Get Started Button'
+    );
   } );
 
   // Fire an event when Add a School is cancelled
   $( '#introduction .add-cancel' ).click( function() {
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'School Cost Comparison',
-      label: 'Cancel Button'
-    } );
+    track( 'School Interactions', 'School Cost Comparison', 'Cancel Button' );
   } );
 
   // Fire an event when Continue is clicked
   $( '#introduction .continue' ).click( function() {
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'School Cost Comparison',
-      label: 'Continue Button'
-    } );
-    console.log( '#introduction .continue clicked' );
+    track( 'School Interactions', 'School Cost Comparison', 'Continue Button' );
+    analyticsLog( '#introduction .continue clicked' );
   } );
 
   // Fire an event when Add another school is clicked
   $( '#introduction .add-another-school' ).click( function() {
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'School Cost Comparison',
-      label: 'Add another school Button'
-    } );
+    track(
+      'School Interactions',
+      'School Cost Comparison',
+      'Add another school Button'
+    );
   } );
 
   // Fire an event when adding a school.
@@ -241,58 +170,26 @@ const PFCAnalytics = ( function() {
     if ( $( '#finaidoffer' ).is( ':checked' ) ) {
       offer = 'Yes';
     }
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'Total Schools Added',
-      label: schoolCount
-    } );
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'School Added',
-      label: schoolID
-    } );
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'Program Type',
-      label: program
-    } );
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'Program Length',
-      label: prgmlength
-    } );
-
+    track( 'School Interactions', 'Total Schools Added', schoolCount );
+    track( 'School Interactions', 'School Added', schoolID );
+    track( 'School Interactions', 'Program Type', program );
+    track( 'School Interactions', 'Program Length', prgmlength );
 
     if ( offer === 'Yes' ) {
+      // TODO: Determine what the label should be for this?
       window.dataLayer.push( {
         event: 'School Interactions',
         action: 'Financial Aid Clicked'
       } );
       if ( $( '#xml-text' ).val() === '' && kbyoss === 'Yes' ) {
-        window.dataLayer.push( {
-          event: 'School Interactions',
-          action: 'School Added - XML',
-          label: 'Blank'
-        } );
+        track( 'School Interactions', 'School Added - XML', 'Blank' );
       } else if ( $( '#xml-text' ).val() !== '' && kbyoss === 'Yes' ) {
-        window.dataLayer.push( {
-          event: 'School Interactions',
-          action: 'School Added - XML',
-          label: 'XML text'
-        } );
+        track( 'School Interactions', 'School Added - XML', 'XML text' );
       }
     } else {
-      window.dataLayer.push( {
-        event: 'School Interactions',
-        action: 'Housing',
-        label: housing
-      } );
+      track( 'School Interactions', 'Housing', housing );
       if ( control === 'Public' ) {
-        window.dataLayer.push( {
-          event: 'School Interactions',
-          action: 'Residency',
-          label: residency
-        } );
+        track( 'School Interactions', 'Residency', residency );
       }
     }
   }
@@ -312,11 +209,7 @@ const PFCAnalytics = ( function() {
   // Fire event when user clicks the arrows to open sections
   $( '.arrw-collapse' ).click( function() {
     const arrwName = $( this ).attr( 'data-arrwname' );
-    window.dataLayer.push( {
-      event: 'School Interactions',
-      action: 'Drop Down',
-      label: arrwName
-    } );
+    track( 'School Interactions', 'Drop Down', arrwName );
   } );
 
 } )( $ );
