@@ -14,6 +14,7 @@ const COMMON_BUNDLE_NAME = 'common.js';
 /* Set warnings to true to show linter-style warnings.
    Set mangle to false and beautify to true to debug the output code. */
 const COMMON_UGLIFY_CONFIG = new UglifyWebpackPlugin( {
+  cache: true,
   parallel: true,
   uglifyOptions: {
     ie8: false,
@@ -43,7 +44,7 @@ const COMMON_MODULE_CONFIG = {
           targets: {
             browsers: BROWSER_LIST.LAST_2_IE_9_UP
           },
-          debug: true
+          debug: false
         } ] ]
       }
     }
@@ -58,6 +59,12 @@ const COMMON_MODULE_CONFIG = {
 const COMMON_CHUNK_CONFIG = new webpack.optimize.SplitChunksPlugin( {
   name: COMMON_BUNDLE_NAME
 } );
+
+const STATS_CONFIG  = {
+  stats: {
+    entrypoints: false
+  }
+};
 
 const conf = {
   cache: true,
@@ -78,7 +85,8 @@ const conf = {
   plugins: [
     COMMON_CHUNK_CONFIG,
     COMMON_UGLIFY_CONFIG
-  ]
+  ],
+  ...STATS_CONFIG
 };
 
 
