@@ -32,7 +32,7 @@ const COMMON_MODULE_CONFIG = {
           targets: {
             browsers: BROWSER_LIST.LAST_2_IE_9_UP
           },
-          debug: true
+          debug: false
         } ] ]
       }
     }
@@ -42,6 +42,7 @@ const COMMON_MODULE_CONFIG = {
 /* Set warnings to true to show linter-style warnings.
    Set mangle to false and beautify to true to debug the output code. */
 const COMMON_UGLIFY_CONFIG = new UglifyWebpackPlugin( {
+  cache: true,
   parallel: true,
   uglifyOptions: {
     ie8: false,
@@ -60,8 +61,14 @@ const COMMON_CHUNK_CONFIG = new webpack.optimize.SplitChunksPlugin( {
   name: COMMON_BUNDLE_NAME
 } );
 
+const STATS_CONFIG = {
+  stats: {
+    entrypoints: false
+  }
+};
 
 const commonConf = {
+  cache: true,
   module: COMMON_MODULE_CONFIG,
   mode: 'production',
   output: {
@@ -69,10 +76,12 @@ const commonConf = {
   },
   plugins: [
     COMMON_UGLIFY_CONFIG
-  ]
+  ],
+  ...STATS_CONFIG
 };
 
 const externalConf = {
+  cache: true,
   module: COMMON_MODULE_CONFIG,
   mode: 'production',
   output: {
@@ -80,7 +89,8 @@ const externalConf = {
   },
   plugins: [
     COMMON_UGLIFY_CONFIG
-  ]
+  ],
+  ...STATS_CONFIG
 };
 
 const modernConf = {
@@ -93,7 +103,8 @@ const modernConf = {
   plugins: [
     COMMON_CHUNK_CONFIG,
     COMMON_UGLIFY_CONFIG
-  ]
+  ],
+  ...STATS_CONFIG
 };
 
 const onDemandHeaderRawConf = {
@@ -111,10 +122,12 @@ const appsConf = {
   plugins: [
     COMMON_CHUNK_CONFIG,
     COMMON_UGLIFY_CONFIG
-  ]
+  ],
+  ...STATS_CONFIG
 };
 
 const spanishConf = {
+  cache: true,
   module: COMMON_MODULE_CONFIG,
   mode: 'production',
   output: {
@@ -122,7 +135,8 @@ const spanishConf = {
   },
   plugins: [
     COMMON_UGLIFY_CONFIG
-  ]
+  ],
+  ...STATS_CONFIG
 };
 
 const devConf = {
