@@ -1,19 +1,18 @@
-const configClean = require( '../config' ).clean;
 const del = require( 'del' );
 const gulp = require( 'gulp' );
+const paths = require( '../../config/environment' ).paths;
 
+// Clean CSS out of /cfgov/static_built/css/
+gulp.task( 'clean:css', cb => del( [
+  paths.processed + '/css/**/*',
+  paths.processed + '/apps/**/css/*'
+] ) );
 
-gulp.task( 'clean:css', () => {
-  // Clean CSS out of /cfgov/static_built/css/
-  del( configClean.css + '/**/*' );
-} );
+// Clean JavaScript out of /cfgov/static_built/js/
+gulp.task( 'clean:js', cb => del( [
+  paths.processed + '/js/**/*',
+  paths.processed + '/apps/**/js/*'
+] ) );
 
-gulp.task( 'clean:js', () => {
-  // Clean JavaScript out of /cfgov/static_built/js/
-  del( configClean.js + '/**/*' );
-} );
-
-gulp.task( 'clean', () => {
-  // Clean everything out of /cfgov/static_built/
-  del( configClean.dest + '/**/*' );
-} );
+// Clean everything out of /cfgov/static_built/
+gulp.task( 'clean', () => del( paths.processed + '/**/*' ) );
