@@ -49,4 +49,9 @@ function imagesGeneral() {
 
 gulp.task( 'images:apps', imagesApps );
 gulp.task( 'images:general', imagesGeneral );
-gulp.task( 'images', [ 'images:apps', 'images:general' ] );
+gulp.task( 'images',
+  gulp.parallel(
+    'images:apps',
+    'images:general'
+  )
+);
