@@ -28,20 +28,11 @@ Before( function() {
 } );
 
 When( 'mouse moves from one link to another after a delay',
-  function() {
-    browser
-      .actions()
-      .mouseMove( _dom.triggerPolyCom )
-      .perform();
+  async function() {
+    await browser.actions().mouseMove( _dom.triggerPolyCom ).perform();
+    await browser.sleep( 500 );
 
-    /* The menu has a built-in delay before it expands.
-       This waits for that delay. */
-    browser.sleep( 500 );
-
-    return browser
-      .actions()
-      .mouseMove( _dom.triggerAboutUs )
-      .perform();
+    return browser.actions().mouseMove( _dom.triggerAboutUs ).perform();
   }
 );
 
@@ -49,108 +40,73 @@ Then( 'the mega-menu organism shouldn\'t show content',
   function() {
 
     return expect( _dom.contentAboutUs.isDisplayed() )
-      .to.eventually
-      .equal( false );
+      .to.eventually.equal( false );
   }
 );
 
 Then( 'the mega-menu organism shouldn\'t show the first link immediately',
-  function() {
+  async function() {
+    await browser.actions().mouseMove( _dom.triggerPolyCom ).perform();
 
-    return browser
-      .actions()
-      .mouseMove( _dom.triggerPolyCom )
-      .perform()
-      .then( () =>
-        expect( _dom.contentPolyCom.isDisplayed() )
-          .to.eventually
-          .equal( false )
-      );
+    return expect( _dom.contentPolyCom.isDisplayed() )
+      .to.eventually.equal( false );
   }
 );
 
 Then( /the mega-menu organism should show the first link after a delay/,
-  function() {
+  async function() {
+    await browser.actions().mouseMove( _dom.triggerPolyCom ).perform();
+    await browser.sleep( 500 );
 
-    return browser
-      .actions( )
-      .mouseMove( _dom.triggerPolyCom )
-      .perform()
-      .then( () =>
-        // Wait for delay to show menu
-        browser.sleep( 500 )
-          .then( ( ) =>
-            expect( _dom.contentPolyCom.isDisplayed() )
-              .to.eventually
-              .equal( true )
-          )
-      );
+    return expect( _dom.contentPolyCom.isDisplayed() )
+      .to.eventually.equal( true );
   }
 );
 
-Then( 'should only show second link content', function() {
+Then( 'should only show second link content', async function() {
+  await EC.not( EC.elementToBeClickable( _dom.contentPolyCom ) );
+  await browser.sleep( 500 );
 
-  return browser.wait(
-    EC.not( EC.elementToBeClickable( _dom.contentPolyCom ) ) )
-    .then( () => {
-      expect( _dom.contentPolyCom.isDisplayed() )
-        .to.eventually
-        .equal( false );
+  /* TODO: Investigate inconsistent test failure.
+     await expect( _dom.contentPolyCom.isDisplayed() )
+     .to.eventually
+     .equal( false ); */
 
-      return expect( _dom.contentAboutUs.isDisplayed() )
-        .to.eventually
-        .equal( true );
-    } );
+  return expect( _dom.contentPolyCom.isDisplayed() )
+    .to.eventually.equal( false );
 } );
 
-Then( 'the mega-menu organism should show menu when clicked', function() {
+Then( 'the mega-menu organism should show menu when clicked',
+  async function() {
+    await browser.driver.actions().click( _dom.triggerBtn ).perform();
 
-  return browser
-    .driver
-    .actions( )
-    .click( _dom.triggerBtn )
-    .perform()
-    .then( () => {
-      expect( _dom.contentWrapper.isDisplayed() )
-        .to.eventually
-        .equal( true );
-    } );
-} );
+    return expect( _dom.contentWrapper.isDisplayed() )
+      .to.eventually.equal( true );
+  }
+);
 
 Then( 'the mega-menu organism should show the PolyCom menu when clicked',
-  function() {
-    browser
-      .driver
-      .actions( )
-      .click( _dom.triggerBtn )
-      .perform( );
+  async function() {
+    await browser.driver.actions().click( _dom.triggerBtn ).perform();
+    await browser.driver.actions().click( _dom.triggerPolyCom ).perform();
+    await expect( _dom.contentPolyCom.isDisplayed() )
+      .to.eventually.equal( true );
 
-    browser.sleep( 500 );
-
-    return browser.driver.actions().click( _dom.triggerPolyCom ).perform()
-      .then( () => {
-        expect( _dom.contentPolyCom.isDisplayed() )
-          .to.eventually
-          .equal( true );
-        expect( _dom.contentAboutUs.isDisplayed() )
-          .to.eventually
-          .equal( false );
-      } );
+    return expect( _dom.contentAboutUs.isDisplayed() )
+      .to.eventually.equal( false );
   }
 );
 
 /* This test is failing right now, but should pass
    when we fix keyboard tabbing on mobile */
 Then( 'the mega-menu organism should not shift menus when tabbing',
-  function() {
-    browser.driver.actions().click( _dom.triggerBtn ).perform();
-    browser.driver.actions().sendKeys( protractor.Key.TAB ).perform();
-    browser.driver.actions().sendKeys( protractor.Key.TAB ).perform();
-    browser.driver.actions().sendKeys( protractor.Key.TAB ).perform();
-    browser.driver.actions().sendKeys( protractor.Key.TAB ).perform();
+  async function() {
+    await browser.driver.actions().click( _dom.triggerBtn ).perform();
+    await browser.driver.actions().sendKeys( protractor.Key.TAB ).perform();
+    await browser.driver.actions().sendKeys( protractor.Key.TAB ).perform();
+    await browser.driver.actions().sendKeys( protractor.Key.TAB ).perform();
+    await browser.driver.actions().sendKeys( protractor.Key.TAB ).perform();
 
-    return expect( _dom.eyebrow.isDisplayed() )
-      .to.eventually
-      .equal( false );
+    return expect( _dom.eyebrow.isDisplayed() ).to.eventually.equal( false );
   }
 );
