@@ -1,27 +1,29 @@
 const util = require( '../../util/index.js' );
-const shouldShouldnt = util.shouldShouldnt;
+const isShould = util.isShould;
 
 const { When, Then } = require( 'cucumber' );
 const { expect } = require( 'chai' );
 
 When( 'there exists a reusable text snippet on this page',
-  function() {
-    const rts = element( by.css( '.m-reusable-text-snippet' ) );
+  async function() {
+    const rts = await element( by.css( '.m-reusable-text-snippet' ) );
+
     return expect( rts.isPresent() ).to.eventually.equal( true );
   }
 );
 
 Then( /the snippet output (should|shouldn't) include a slug-style header/,
-  function( includeHeader ) {
+  async function( includeHeader ) {
     let rts;
 
-    if ( shouldShouldnt( includeHeader ) ) {
-      rts = element.all( by.css( '.m-reusable-text-snippet' ) ).first();
+    if ( isShould( includeHeader ) ) {
+      rts = await element.all( by.css( '.m-reusable-text-snippet' ) ).first();
     } else {
-      rts = element.all( by.css( '.m-reusable-text-snippet' ) ).last();
+      rts = await element.all( by.css( '.m-reusable-text-snippet' ) ).last();
     }
 
-    return expect( rts.element( by.css( '.m-slug-header' ) ).isPresent() )
-      .to.eventually.equal( shouldShouldnt( includeHeader ) );
+    return expect( rts.element( by.css( '.m-slug-header' ) )
+      .isPresent() )
+      .to.eventually.equal( isShould( includeHeader ) );
   }
 );
