@@ -25,7 +25,9 @@ When( /the JS bundles shouldn't contain double arrows or constants/, () => {
 
           const isFailure = transpileRegex.test( contents );
           if ( isFailure ) {
-            reject( new Error( directoryMapKeys[i] + ' contains const or ()=>' ) );
+            reject(
+              new Error( directoryMapKeys[i] + ' contains const or ()=>' )
+            );
           } else if ( i === directoryMapLength - 1 ) {
             // All files searched.
             resolve();
