@@ -2,21 +2,21 @@ const BasePage = require( './base-page.js' );
 
 class BaseFilterablePage extends BasePage {
 
-  constructor() {
-    super();
-    this.results = element.all( by.css( '.o-post-preview_content' ) );
-    this.firstResult = this.results.first();
-    this.lastResult = this.results.last();
-    this.multiselect = element.all( by.css( '.cf-multi-select' ) );
+  async setElements() {
+    this.results = await element.all( by.css( '.o-post-preview_content' ) );
+    this.firstResult = this.results[0];
+    this.lastResult = this.results[this.results.length - 1];
+    this.multiselect = await element.all( by.css( '.cf-multi-select' ) );
   }
 
   getResultsCount() {
-    return this.results.count()
-      .then( count => count.toString() );
+    const count = this.results.length;
+
+    return count.toString();
   }
 
-  getResultText( resultPosition = 'first' ) {
-    return this[resultPosition + 'Result'].getText();
+  async getResultText( resultPosition = 'first' ) {
+    return await this[resultPosition + 'Result'].getText();
   }
 }
 
