@@ -1,6 +1,5 @@
 const _oFilterableListControls =
   element( by.css( '.o-filterable-list-controls' ) );
-
 const multiselect = require( '../shared_objects/multi-select' );
 const EC = protractor.ExpectedConditions;
 
@@ -9,13 +8,11 @@ function _getFilterableElement( selector ) {
   return _oFilterableListControls.element( by.css( selector ) );
 }
 
-function open() {
+async function open() {
   const expandable = this.mExpandable;
+  await expandable.click();
 
-  return expandable.click()
-    .then( function() {
-      return browser.wait( EC.elementToBeClickable( expandable ) );
-    } );
+  return browser.wait( EC.elementToBeClickable( expandable ) );
 }
 
 function close() {
