@@ -7,19 +7,23 @@ const stayInformedSection = {
   stayInformedSection: _stayInformedSection,
 
   stayInformedSectionTitle:
-  _stayInformedSection.element( by.css( '.m-slug-header .a-heading' ) ),
+    _stayInformedSection.element(
+      by.css( '.m-slug-header .a-heading' )
+    ),
 
   emailSubscribeForm: _emailSubscribeForm,
 
   emailFormInput:
-  _emailSubscribeForm.element( by.css( 'input[type="email"]' ) ),
+    _emailSubscribeForm.element( by.css( 'input[type="email"]' ) ),
 
-  emailFormLabel: _emailSubscribeForm.element( by.css( 'label' ) ),
+  emailFormLabel:
+    _emailSubscribeForm.element( by.css( 'label' ) ),
 
   emailFormHiddenField:
-  _emailSubscribeForm.element( by.css( 'input[type="hidden"]' ) ),
+    _emailSubscribeForm.element( by.css( 'input[type="hidden"]' ) ),
 
-  emailFormBtn: _emailSubscribeForm.element( by.css( 'input[type="submit"]' ) )
+  emailFormBtn:
+    _emailSubscribeForm.element( by.css( 'input[type="submit"]' ) )
 
 };
 
